HBP-87. Move utils under tvb_multiscale/core and remove tvb_utils separate module.
