+++ tvb_multiscale/core/config.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import sys
 import os
 import numpy
 import inspect
@@ -14,7 +13,7 @@ from tvb.datatypes import cortex, connectivity
 
 import tvb_data
 
-from tvb_utils.log_utils import initialize_logger as initialize_logger_base
+from tvb_multiscale.core.log_utils import initialize_logger as initialize_logger_base
 from tvb.contrib.scripts.utils.file_utils import safe_makedirs
 
 TVB_NEST_DIR = os.path.abspath(__file__).split("tvb_multiscale")[0]
